cover edge cases, make game creation & join faster

+++ src/main/java/ch/uzh/ifi/hase/soprafs22/service/GameService.java
@@ -122,6 +122,7 @@ public class GameService {
                 gameToLeave.getPlayerIds().remove(playerId);
                 List<String> currentPlayerNames=gameToLeave.getPlayerNames();
                 Player playerToDelete=playerRepository.findByPlayerId(playerId);
+                if (playerToDelete == null) return;
                 currentPlayerNames.remove(playerToDelete.getPlayerName());
                 gameToLeave.setPlayerNames(currentPlayerNames);
                 playerRepository.delete(playerToDelete);
@@ -133,7 +134,7 @@ public class GameService {
     }
 
     // function for joining a game
-    public Game joinGame(Long gameId, String token){
+    public synchronized Game joinGame(Long gameId, String token) {
         Game game = this.getGame(gameId);
         User userToJoin = userRepository.findByToken(token);
         if (userToJoin.getStatus() == UserStatus.ONLINE) {
@@ -154,7 +155,7 @@ public class GameService {
         } else { throw new ResponseStatusException(HttpStatus.CONFLICT, "User is not logged in, cannot join a game!"); }
     }
 
-   private void startGame(Game game) {
+   private synchronized void startGame(Game game) {
        // get not played white cards
         List <Card> whiteCards=cardRepository.findByDeckIdAndIsWhiteAndIsPlayed(game.getDeckID(),true,false);
         int upperbound=whiteCards.size();
@@ -172,6 +173,7 @@ public class GameService {
            unique.add(number); // adding to the set
        }
        int card_Index=0;
+       List<Card> cardsToSave = new ArrayList<Card>();
        for(Long playerId: game.getPlayerIds()){
            Player currentPlayer=playerRepository.findByPlayerId(playerId);
            for(int i=0; i<10; i++){
@@ -180,13 +182,16 @@ public class GameService {
                playerCards.add(currentCard);
                currentPlayer.setCardsOnHands(playerCards);
                currentCard.setPlayed(true);
-               cardRepository.saveAndFlush(currentCard);
+               cardsToSave.add(currentCard);
+//               cardRepository.saveAndFlush(currentCard);
                card_Index++;
            }
            playerRepository.saveAndFlush(currentPlayer);
        }
+       cardRepository.saveAll(cardsToSave);
+       cardRepository.flush();
 
-        GameRound currentGameRound=gameRoundService.startNewRound(game);
+       GameRound currentGameRound=gameRoundService.startNewRound(game);
        return;
     }
 
@@ -202,7 +207,7 @@ public class GameService {
     }
 
     // method if a player decides to leave the game
-    public void leaveGame(Long gameId, String token) { //frontend knows the winner -->
+    public synchronized void leaveGame(Long gameId, String token) { //frontend knows the winner -->
         Game gameToLeave = this.getGame(gameId);
         User userToRemove = userRepository.findByToken(token);
         Player playerToRemove = playerRepository.findByPlayerId(userToRemove.getUserId());   // from DIEGO
@@ -218,17 +223,13 @@ public class GameService {
         removePlayerFromGame(gameToLeave, userToRemove);
         // if it is the last player, also delete the game/gameRounds/
         if (lastPlayer) {
-            //TODO delete cards/deck after game is finished
-            //TODO delete gameRound/game
-            //deleteGame(gameToLeave); //--> call this function and delete the below 2 lines
+            // DIEGO: security, if not all players properly left the game, delete them
+            for (Long id : gameToLeave.getPlayerIds()) {
+                playerRepository.deleteById(id);
+                playerRepository.flush();
+            }
             gameRepository.delete(gameToLeave);
             gameRepository.flush();
-//            Deck deckToBeDeleted = deckRepository.findByDeckId(gameToLeave.getDeckID());
-//            List<Card> cardsToBeDeleted = cardRepository.findByDeckId(gameToLeave.getDeckID());
-//            cardRepository.deleteAll(cardsToBeDeleted);
-//            cardRepository.flush();
-//            deckRepository.delete(deckToBeDeleted);
-//            deckRepository.flush();
         }
     }
